refactor(session-activity): 外移 activity 面板纯 helper 层
- 2679→2086 行：activityPanelHelpers.ts 592 行（types+常量+~40 纯函数）
- renderSemanticSummaryList 等被 aria 测试锁的签名未动
- baseline 重生成 + 新文件登记豁免
验证：tsc 本文件族零 error；session-activity 15 文件 202 passed
