+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T19:24:47.279Z",
+  "generatedAt": "2026-08-30T20:27:22.271Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
@@ -14,28 +14,8 @@
       "thresholdSource": "policy"
     },
     {
-      "path": "src-tauri/src/bin/cc_gui_daemon/daemon_state.rs",
-      "lines": 5339,
-      "type": "rust",
-      "policyId": "default-source",
-      "priority": "P1",
-      "warnThreshold": 2600,
-      "failThreshold": 3000,
-      "thresholdSource": "policy"
-    },
-    {
       "path": "src/features/git-history/components/git-history-panel/components/GitHistoryPanelImpl.tsx",
       "lines": 4933,
-      "type": "ts/tsx",
-      "policyId": "feature-hotpath",
-      "priority": "P1",
-      "warnThreshold": 2400,
-      "failThreshold": 2800,
-      "thresholdSource": "policy"
-    },
-    {
-      "path": "src/features/threads/hooks/useThreadMessaging.ts",
-      "lines": 4600,
       "type": "ts/tsx",
       "policyId": "feature-hotpath",
       "priority": "P1",
@@ -49,6 +29,16 @@
       "type": "ts/tsx",
       "policyId": "test-files",
       "priority": "P2",
+      "warnThreshold": 2600,
+      "failThreshold": 3000,
+      "thresholdSource": "policy"
+    },
+    {
+      "path": "src/features/layout/hooks/useLayoutNodes.tsx",
+      "lines": 4127,
+      "type": "ts/tsx",
+      "policyId": "default-source",
+      "priority": "P1",
       "warnThreshold": 2600,
       "failThreshold": 3000,
       "thresholdSource": "policy"
@@ -70,6 +60,16 @@
       "policyId": "styles",
       "priority": "P1",
       "warnThreshold": 2200,
+      "failThreshold": 2800,
+      "thresholdSource": "policy"
+    },
+    {
+      "path": "src/features/threads/hooks/useThreadMessaging.ts",
+      "lines": 4025,
+      "type": "ts/tsx",
+      "policyId": "feature-hotpath",
+      "priority": "P1",
+      "warnThreshold": 2400,
       "failThreshold": 2800,
       "thresholdSource": "policy"
     },
@@ -204,16 +204,6 @@
       "thresholdSource": "policy"
     },
     {
-      "path": "src/features/layout/hooks/useLayoutNodes.tsx",
-      "lines": 3235,
-      "type": "ts/tsx",
-      "policyId": "default-source",
-      "priority": "P1",
-      "warnThreshold": 2600,
-      "failThreshold": 3000,
-      "thresholdSource": "policy"
-    },
-    {
       "path": "src-tauri/src/browser_agent/mod.rs",
       "lines": 3224,
       "type": "rust",
@@ -244,7 +234,7 @@
       "thresholdSource": "policy"
     },
     {
-      "path": "src-tauri/src/bin/cc_gui_daemon/git.rs",
+      "path": "src-tauri/src/bin/cc_gui_daemon/daemon_state/git.rs",
       "lines": 3170,
       "type": "rust",
       "policyId": "default-source",
@@ -310,16 +300,6 @@
       "policyId": "styles",
       "priority": "P1",
       "warnThreshold": 2200,
-      "failThreshold": 2800,
-      "thresholdSource": "policy"
-    },
-    {
-      "path": "src/features/threads/hooks/useThreadActions.ts",
-      "lines": 3066,
-      "type": "ts/tsx",
-      "policyId": "feature-hotpath",
-      "priority": "P1",
-      "warnThreshold": 2400,
       "failThreshold": 2800,
       "thresholdSource": "policy"
     },
